refactor: recommendation 도메인 계층을 최상위로 승격
