+++ src/main/java/com/techfork/domain/recommendation/repository/RecommendationHistoryRepository.java
@@ -1,6 +1,6 @@
 package com.techfork.domain.recommendation.repository;
 
-import com.techfork.domain.recommendation.entity.RecommendationHistory;
+import com.techfork.recommendation.domain.RecommendationHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecommendationHistoryRepository extends JpaRepository<RecommendationHistory, Long> {
