+++ src/main/java/com/techfork/domain/recommendation/service/LlmRecommendationService.java
@@ -7,8 +7,8 @@ import co.elastic.clients.elasticsearch.core.SearchResponse;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import com.techfork.activity.readpost.application.query.lookup.ReadPostLookupService;
 import com.techfork.domain.recommendation.config.RecommendationProperties;
-import com.techfork.domain.recommendation.entity.RecommendationHistory;
-import com.techfork.domain.recommendation.entity.RecommendedPost;
+import com.techfork.recommendation.domain.RecommendationHistory;
+import com.techfork.recommendation.domain.RecommendedPost;
 import com.techfork.domain.recommendation.repository.RecommendationHistoryRepository;
 import com.techfork.domain.recommendation.repository.RecommendedPostRepository;
 import com.techfork.domain.recommendation.service.MmrService.MmrCandidate;
