+++ src/test/java/com/techfork/domain/recommendation/converter/RecommendationConverterTest.java
@@ -2,7 +2,7 @@ package com.techfork.domain.recommendation.converter;
 
 import com.techfork.post.domain.Post;
 import com.techfork.domain.recommendation.dto.RecommendedPostDto;
-import com.techfork.domain.recommendation.entity.RecommendedPost;
+import com.techfork.recommendation.domain.RecommendedPost;
 import com.techfork.domain.source.entity.TechBlog;
 import com.techfork.useraccount.domain.User;
 import com.techfork.global.util.CloudflareThirdPartyThumbnailOptimizer;
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 import java.time.LocalDateTime;
 
 import static com.techfork.post.fixture.PostFixture.createPost;
-import static com.techfork.domain.recommendation.fixture.RecommendedPostFixture.recommendedPost;
+import static com.techfork.recommendation.fixture.RecommendedPostFixture.recommendedPost;
 import static com.techfork.domain.source.fixture.TechBlogFixture.createTechBlog;
 import static com.techfork.useraccount.fixture.UserFixture.socialUser;
 import static org.assertj.core.api.Assertions.assertThat;
