+++ src/test/java/com/techfork/recommendation/domain/RecommendationHistoryTest.java
@@ -1,4 +1,4 @@
-package com.techfork.domain.recommendation.entity;
+package com.techfork.recommendation.domain;
 
 import com.techfork.post.domain.Post;
 import com.techfork.useraccount.domain.User;
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import java.time.LocalDateTime;
 
 import static com.techfork.post.fixture.PostFixture.createPost;
-import static com.techfork.domain.recommendation.fixture.RecommendedPostFixture.recommendedPost;
+import static com.techfork.recommendation.fixture.RecommendedPostFixture.recommendedPost;
 import static com.techfork.domain.source.fixture.TechBlogFixture.createTechBlog;
 import static com.techfork.useraccount.fixture.UserFixture.socialUser;
 import static org.assertj.core.api.Assertions.assertThat;
